config: the cross-reflector conflict rule in its own module
