+++ src/config.rs
@@ -9,19 +9,20 @@
 //! `raw`, the environment parser in `env`. Each value type pairs `FromStr` with a
 //! matching `Deserialize`, so one validation serves both the TOML path (serde,
 //! located errors) and the environment path (`FromStr`, variable-named errors).
-//! Cross-field and cross-reflector rules live in the `TryFrom` conversions here;
-//! sources are combined in [`Config::from_sources`].
+//! Cross-field rules live in the `TryFrom` conversions here, the cross-reflector ones in
+//! `conflict`; sources are combined in [`Config::from_sources`].
 //!
 //! Reflectors nest under `[reflectors.<name>]` rather than top-level tables to keep
 //! the deserializer off `#[serde(flatten)]`, which would discard the line/column of
 //! every value error.
 
+mod conflict;
 mod env;
 mod error;
 mod raw;
 mod value;
 
-pub(crate) use self::error::{ConfigError, Protocol};
+pub(crate) use self::error::ConfigError;
 pub(crate) use self::value::{
     AddressFamily, GroupList, InterfaceName, LogLevel, PeerList, PortList, ReflectorName,
 };
@@ -34,13 +35,9 @@ use std::time::Duration;
 
 use serde::Deserialize;
 
+use self::conflict::check_conflicts;
 use self::raw::{RawConfig, RawReflector};
 use crate::net::mac::MacSet;
-use crate::net::mdns::{MDNS_GROUP_V4, MDNS_GROUP_V6, MDNS_PORT};
-use crate::net::ssdp::{
-    SSDP_GROUP_V4, SSDP_GROUP_V6_LINK_LOCAL, SSDP_GROUP_V6_SITE_LOCAL, SSDP_PORT,
-};
-use crate::net::wsd::{WSD_GROUP_V4, WSD_GROUP_V6, WSD_PORT};
 
 /// Wake-on-LAN settings (present only when `WoL` is enabled for the reflector).
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -59,55 +56,6 @@ impl Wol {
     }
 }
 
-/// The destinations a protocol's filter admits on one port.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum Reach {
-    /// Any address the family uses: the Wake-on-LAN filter pins the port alone.
-    Any(AddressFamily),
-    Group(IpAddr),
-    /// The limited broadcast and the segment's directed one.
-    Broadcast,
-}
-
-impl Reach {
-    /// Whether a datagram exists that both admit.
-    fn overlaps(self, other: Reach) -> bool {
-        match (self, other) {
-            (Self::Any(a), Self::Any(b)) => families_overlap(a, b),
-            (Self::Any(family), Self::Group(group)) | (Self::Group(group), Self::Any(family)) => {
-                family.uses(group)
-            }
-            (Self::Any(family), Self::Broadcast) | (Self::Broadcast, Self::Any(family)) => {
-                family.uses_ipv4()
-            }
-            (Self::Group(a), Self::Group(b)) => a == b,
-            (Self::Broadcast, Self::Broadcast) => true,
-            (Self::Group(_), Self::Broadcast) | (Self::Broadcast, Self::Group(_)) => false,
-        }
-    }
-}
-
-/// A class of datagrams an entry relays: those to `port` and `reach` arriving on `ingress`,
-/// re-emitted on `egress`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct Flow<'a> {
-    protocol: Protocol,
-    ingress: &'a InterfaceName,
-    egress: &'a InterfaceName,
-    port: u16,
-    reach: Reach,
-}
-
-impl Flow<'_> {
-    /// Whether a datagram exists in both flows.
-    fn overlaps(&self, other: &Flow<'_>) -> bool {
-        self.ingress == other.ingress
-            && self.egress == other.egress
-            && self.port == other.port
-            && self.reach.overlaps(other.reach)
-    }
-}
-
 /// The transparent UDP relay's settings (present only when `udp_ports` is set).
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) struct UdpRelay {
@@ -117,20 +65,6 @@ pub(crate) struct UdpRelay {
     pub(crate) groups: Option<GroupList>,
     /// Whether broadcasts on those ports are relayed too.
     pub(crate) broadcast: bool,
-}
-
-impl UdpRelay {
-    fn destinations(&self) -> Vec<(u16, Reach)> {
-        let groups = self.groups.as_deref().unwrap_or(&[]);
-        self.ports
-            .iter()
-            .map(|port| port.get())
-            .flat_map(|port| {
-                let groups = groups.iter().map(move |group| (port, Reach::Group(*group)));
-                groups.chain(self.broadcast.then_some((port, Reach::Broadcast)))
-            })
-            .collect()
-    }
 }
 
 /// SSDP settings (present only when SSDP is enabled for the reflector).
@@ -183,134 +117,6 @@ impl Reflector {
             target_peers: self.source_peers.clone(),
             ..self.clone()
         }
-    }
-
-    /// The `(source, target)` pairs this entry relays over: its own, plus the reverse when
-    /// bidirectional.
-    fn directions(&self) -> impl Iterator<Item = (&InterfaceName, &InterfaceName)> {
-        std::iter::once((&self.source_if, &self.target_if)).chain(
-            self.bidirectional
-                .then_some((&self.target_if, &self.source_if)),
-        )
-    }
-
-    /// Every flow the entry's protocols relay. mDNS, SSDP and WSD capture on both interfaces
-    /// whatever the entry's direction, queries on the source and responses on the target, for
-    /// the groups of the families the entry uses; Wake-on-LAN admits anything on its ports and
-    /// the relay its groups and broadcast, each on the entry's directions.
-    fn flows(&self) -> Vec<Flow<'_>> {
-        let family = self.address_family;
-        let mut flows = Vec::new();
-        let mut discovery = |protocol, port, groups: &[IpAddr]| {
-            for group in groups.iter().filter(|group| family.uses(**group)) {
-                let legs = [
-                    (&self.source_if, &self.target_if),
-                    (&self.target_if, &self.source_if),
-                ];
-                flows.extend(legs.map(|(ingress, egress)| Flow {
-                    protocol,
-                    ingress,
-                    egress,
-                    port,
-                    reach: Reach::Group(*group),
-                }));
-            }
-        };
-        if self.mdns {
-            let v4 = IpAddr::V4(MDNS_GROUP_V4);
-            let v6 = IpAddr::V6(MDNS_GROUP_V6);
-            discovery(Protocol::Mdns, MDNS_PORT, &[v4, v6]);
-        }
-        if self.ssdp.is_some() {
-            let v4 = IpAddr::V4(SSDP_GROUP_V4);
-            let link_local = IpAddr::V6(SSDP_GROUP_V6_LINK_LOCAL);
-            let site_local = IpAddr::V6(SSDP_GROUP_V6_SITE_LOCAL);
-            discovery(Protocol::Ssdp, SSDP_PORT, &[v4, link_local, site_local]);
-        }
-        if self.wsd {
-            let v4 = IpAddr::V4(WSD_GROUP_V4);
-            let v6 = IpAddr::V6(WSD_GROUP_V6);
-            discovery(Protocol::Wsd, WSD_PORT, &[v4, v6]);
-        }
-        for (ingress, egress) in self.directions() {
-            if let Some(wol) = &self.wol {
-                flows.extend(wol.ports.iter().map(|port| Flow {
-                    protocol: Protocol::Wol,
-                    ingress,
-                    egress,
-                    port: port.get(),
-                    reach: Reach::Any(family),
-                }));
-            }
-            if let Some(udp) = &self.udp {
-                flows.extend(udp.destinations().into_iter().map(|(port, reach)| Flow {
-                    protocol: Protocol::Udp,
-                    ingress,
-                    egress,
-                    port,
-                    reach,
-                }));
-            }
-        }
-        flows
-    }
-
-    /// The protocol on which `self` and `other` would reflect the same packet twice, if any: a
-    /// protocol both enable, or a flow of one the UDP relay of the other overlaps. Two
-    /// discovery protocols on one port don't duplicate: each classifier admits only its own
-    /// messages. The relay admits every datagram it captures, so it duplicates any protocol
-    /// capturing the same on the same leg, and the conflict is named after that protocol.
-    fn conflicts_with(&self, other: &Reflector) -> Option<Protocol> {
-        self.shared_protocol(other).or_else(|| {
-            self.relay_overlap(other.flows())
-                .or_else(|| other.relay_overlap(self.flows()))
-                .map(|(protocol, _)| protocol)
-        })
-    }
-
-    /// A protocol both enable on a shared direction with overlapping MAC selection and address
-    /// family (for `WoL`, also a shared port).
-    fn shared_protocol(&self, other: &Reflector) -> Option<Protocol> {
-        if !self
-            .directions()
-            .any(|mine| other.directions().any(|theirs| theirs == mine))
-        {
-            return None;
-        }
-        if !macs_overlap(self.macs.as_ref(), other.macs.as_ref())
-            || !families_overlap(self.address_family, other.address_family)
-        {
-            return None;
-        }
-        if let (Some(a), Some(b)) = (&self.wol, &other.wol)
-            && a.ports.iter().any(|port| b.ports.contains(port))
-        {
-            return Some(Protocol::Wol);
-        }
-        if self.mdns && other.mdns {
-            return Some(Protocol::Mdns);
-        }
-        if self.ssdp.is_some() && other.ssdp.is_some() {
-            return Some(Protocol::Ssdp);
-        }
-        if self.wsd && other.wsd {
-            return Some(Protocol::Wsd);
-        }
-        None
-    }
-
-    /// The first of `others` whose datagrams the entry's UDP relay would carry a second time,
-    /// as its protocol and port.
-    fn relay_overlap<'a>(
-        &self,
-        others: impl IntoIterator<Item = Flow<'a>>,
-    ) -> Option<(Protocol, u16)> {
-        let mine = self.flows();
-        let relay = || mine.iter().filter(|flow| flow.protocol == Protocol::Udp);
-        others
-            .into_iter()
-            .find(|other| relay().any(|mine| mine.overlaps(other)))
-            .map(|other| (other.protocol, other.port))
     }
 }
 
@@ -450,15 +256,7 @@ impl TryFrom<(String, RawReflector)> for Reflector {
             bidirectional: raw.bidirectional,
             udp,
         };
-        let duplicated = {
-            let flows = reflector.flows();
-            let others = flows
-                .iter()
-                .copied()
-                .filter(|flow| flow.protocol != Protocol::Udp);
-            reflector.relay_overlap(others)
-        };
-        if let Some((protocol, port)) = duplicated {
+        if let Some((protocol, port)) = reflector.relay_duplicates() {
             return Err(ConfigError::UdpRelayDuplicates {
                 name: reflector.name,
                 port,
@@ -639,46 +437,6 @@ fn protocol_list(reflector: &Reflector) -> String {
         ));
     }
     protocols.join(", ")
-}
-
-/// Two MAC selections overlap when they share at least one address, or either is
-/// absent (an absent filter matches any device).
-fn macs_overlap(a: Option<&MacSet>, b: Option<&MacSet>) -> bool {
-    match (a, b) {
-        (Some(a), Some(b)) => a.iter().any(|mac| b.contains(mac)),
-        _ => true,
-    }
-}
-
-/// Two address families overlap when they both carry the same IP version.
-fn families_overlap(a: AddressFamily, b: AddressFamily) -> bool {
-    (a.uses_ipv4() && b.uses_ipv4()) || (a.uses_ipv6() && b.uses_ipv6())
-}
-
-/// Reject any pair of reflectors that share a name or would reflect the same packet twice. Names are the
-/// canonical (lowercased) identity, so `==` catches keys that only differ in case or whitespace — which
-/// `merge_env` folds env-vs-file but the file table cannot.
-fn check_conflicts(reflectors: &[Reflector]) -> Result<(), ConfigError> {
-    for (i, a) in reflectors.iter().enumerate() {
-        for b in &reflectors[i + 1..] {
-            if a.name == b.name {
-                return Err(ConfigError::DuplicateReflectorName {
-                    name: a.name.clone(),
-                });
-            }
-            if let Some(protocol) = a.conflicts_with(b) {
-                return Err(ConfigError::ConflictingReflectors {
-                    protocol,
-                    first: a.name.clone(),
-                    second: b.name.clone(),
-                    source_if: a.source_if.clone(),
-                    target_if: a.target_if.clone(),
-                });
-            }
-        }
-    }
-    log::debug!("no reflector conflicts");
-    Ok(())
 }
 
 #[cfg(test)]
